MacViews: Adds NativeWidgetMac and gets views target compiling on Mac.
This is a minimal amount to get the `views` target compiling on Mac. It
adds an implementation of NativeWidgetMac that spews NOTIMPLEMENTED().
So, nothing works yet, but this gives an FYI builder something to chew
on to catch regressions in things ported so far.

BUG=366007

Review URL: https://codereview.chromium.org/249113002