sync: Refactor autofill integration tests
Refactors the autofill integration tests so they no longer need to rely
on snapshots or self-notifications.

Deletes the integration tests around disabling the type or disabling
sync, since we decided in the review of r269198 that those tests are not
valuable.

Introduces KeysMatchStatusChangeChecker, a fairly typical
MultiClientStatusChangeChecker based on the existing KeysMatch()
function.

Introduces ProfilesMatchStatusChecker, which might be the first
StatusChangeChecker to not register as an observer of the
ProfileSyncService.  Autofill profiles are stored and synced on the DB
thread but cached on the UI thread.  There's no guarantee that the UI
thread cache of autofill profile data would be up to date by the time
the sync cycle notificaiton is emitted.  Instead of listening for sync
notifications, the ProfileMatchStatusChangeChecker registers with the
owner of the UI thread cache of autofill profile data, the
PersonalDataManager, so it can be notified directly when the cache is
updated.

Since it uses a new notification mechanism, there is some reason to
believe this could fix issue 152551.

BUG=97780,95742,152551

Review URL: https://codereview.chromium.org/303643003