WebRTC recipe: Fix sizes step for Chromium builders.
This step is not a normal test but a Python script that
should also be run as a perf test.
It was also added incorrectly to testers - it should only
run on builders. I will take care of migrating the old sizes
data to the new Perf ID for the Windows builder.

A bonus compared to the pre-recipe switch is that we now
also run sizes on the Android builder.

A more thorough attempt to address code duplication is done
in https://codereview.chromium.org/298273004/ but I'd like
to get this in first, since our bots in chromium.webrtc.fyi
are red because of this (and I want to avoid reverting the
recipes switch).

BUG=346198
TEST=None

Review URL: https://codereview.chromium.org/303463007