Use --skipped=always on retries for layout tests.
If one is testing a patch that removes some [ Skip ] lines from the layout
tests' TestExpectations files, and any of those tests fail, when we retry
the tests without the patch, the tests would be run a second time (which
is not what you want, you want the failing tests to be skipped).

This patch fixes that.

R=ojan@chromium.org
BUG=376129

Review URL: https://codereview.chromium.org/302973002