[Telemetry] Cache browser finder.
It turns out we were creating one of these for nearly every unittest. This led
to a couple of problems:

1. Cycle time. With this patch, "tools/telemetry/run_tests Browser" is 20-25%
faster (195s -> 150s).
2. Spam. With this patch, we output many fewer warnings about missing flash.

This is a reland of https://codereview.chromium.org/240593005/ , which broke
the cros bots. This fixes that bug by no longer modifying the copy mechanism.

BUG=337829

Review URL: https://codereview.chromium.org/293133005