Add 'build.dead' cleanup to the proper callback chain.
The cleanup in writeSourcedata() is currently not added to the update
callback chain, so the step doesn't wait for it to finish. This causes
the delete to continue in the background of the next step, which can
cause trouble due to excessive file I/O (e.g. timeouts on slow
hardware).

BUG=377862
R=iannucci@chromium.org, maruel@chromium.org

Review URL: https://codereview.chromium.org/303123005