Introduce the MicButtonTargeter class
Introduce MicButtonTargeter, a derived class of MaskedViewTargeter,
to define the circular hit-test region of MicButton.

We will need to be able to access the hit test mask corresponding
to a view by calling into its installed MaskedViewTargeter, so
GetHitTestMask() should be made public. Furthermore, add a version
of GetEventTargeter() to EventTarget that returns a const pointer.

BUG=377537, 378485
TEST=SpeechViewTest.ClickMicButton

Review URL: https://codereview.chromium.org/302653002