Change instances of s.IsCorruption() to leveldb_env::IsCorruption(s).
LevelDB sometimes returns Status::InvalidArgument() when it encounters
corruption. But LevelDB doesn't provide a Status::IsInvalidArgument()
accessor so we need to indirectly detect it.

This patch revealed that IndexedDBDatabaseOperationTest.CreatePutDelete
had been passing by accident. In order to make it continue to pass, this
patch adds a method to IndexedDBFakeBackingStore.

BUG=322707

Review URL: https://codereview.chromium.org/303073002