Host extensions
This CL introduces HostExtension, an interface for classes that
extend the host with non-core functionality.

Extensions are added to the ChromotingHost. They are used to compile
the list of capabilities reported to the client, which can be used by
the client to determine the availability of the extension.

When a client connects, a HostExtension has the opportunity to create
an HostExtensionSession to hold client/extension state, and to handle
extension messages from that client.

BUG=

NOTRY=true

Review URL: https://codereview.chromium.org/301453003