Fixed bug in dependency recursion in isolate_driver.py.
When run in a component build, isolate_driver recursively descends the
dependency graph to discover implicit shared library dependencies. A
bug in the recursion caused dependencies at certain depths in the
graph to be missed.

Also changed the recursion to use a set to record newly added
dependencies to avoid adding them to the isolate multiple times.
Expanded the ignored set of suffixes slightly.

BUG=378021
TEST=GLES2 conformance tests run via isolate in component build on Mac OS
R=maruel@chromium.org,vadimsh@chromium.org,csharp@chromium.org

Review URL: https://codereview.chromium.org/304993006