Add gapless playback support for AAC playback.
Gapless support for AAC is similar to MP3 but unlike
MP3 which only has padding in the first frame, AAC may
require discarding data across multiple frames.

Unfortunately this means we can't just discard the padding
frames wholesale without changing what's decoded from the
first non-padding frame.

Testing reveals we need to decode the padding frame prior to
the first non-padding frame in order to correctly decode the
non-padding frame.

We accomplish this by introducing the concept of a preroll
buffer which is to be played out prior to the current buffer
but must be entirely discarded.  Luckily this is easily done
with the infrastructure in place for the gapless MP3 playback!

While we need to support post splice buffers with preroll for
gapless playback, for simplicity, pre-splice buffers are not
allowed to have preroll.

In sum, this change:
- Adds SetPrerollBuffer() and preroll_buffer() to the
StreamParserBuffer.
- Refactors splice frame handling into a new
HandleNextBufferWithSpliceFrame() method, introduces
HandleNextBufferWithPreroll()
- Modifies the frame processor to use the frame prior to the
partial append window discard as the preroll.

BUG=371633
TEST=new unittests!
R=wolenetz@chromium.org

Review URL: https://codereview.chromium.org/276573002