mac: Update layout of the autofill element to get contact permissions.
The icon should be left-aligned, to the left of the name. I refactored the
layout code of the autofill popup view to better support both image alignment,
as well as RTL alignment.

This CL depends on https://codereview.chromium.org/286243002/, which has not
yet landed.

See the associated bug for visual confirmation of the correctness of the layout
logic.

BUG=139154

Review URL: https://codereview.chromium.org/296603002