Revert "Fix textfield blur, to prevent resetting the text input client when navigating between text fields."
The fix to use OnTextInputTypeChanged in place of OnFocus and OnBlur to resolve focus loss on text field navigation has introduced a number of regressions. As it turns out, the fix is no longer even necessary to address the original bug that it was intended to address (365741).  This patch is a revert of r270818.

BUG=375125,377359,378914

Review URL: https://codereview.chromium.org/302363002