Fix a lint issue with MediaResourceGetterTest.java.
Currently, the lint warns against using the hardcoded "sdcard" strings. We
should use Environment.getExternalStorageDirectory().getPath() instead of it,
but the path could be inconsistent depending upon the state of the real or
emulated device upon which we are testing. So this CL adds @SuppressLint to
suppress the warnings.

BUG=327768

Review URL: https://codereview.chromium.org/304633003