[Aura] Translate web contents area to nearest physical pixel boundary
Views positions all views, including the views::WebView containing the
aura::Window that contains the content::RenderWidgetHostViewAura, to integral
DIP positions.  With a non-integral device scale factor, this could position
the RWHVA in between physical pixel boundaries which results in filtering the
composited output from the renderer. This looks blurry and bad.

This patch computes the delta to the top-left-most integer physical pixel inside the
desired bounds and translates the web contents layer by that amount. This means that
the top left corner of the web contents shows up at an exact physical pixel
and thus the web contents is not filtered. This means that some amount on the bottom
or right edge of the web contents area may be clipped, but this amount is always less
that one DIP.

BUG=370074, 378620

Review URL: https://codereview.chromium.org/304973003