Pepper: Use an adaptor to invoke FileSystemOperationRunner::OpenFile
PepperFileIOHost keeps a weak pointer of FileSystemOperationRunner so when
it goes away it cannot delete the runner as expected by the fileapi
implementation.

This CL adds an adapter class in the middle so that when OpenFile completes
after PepperFileIOHost destruction, the file is bounced back to another
thread to be closed.

BUG=377599

Review URL: https://codereview.chromium.org/306403005