Fix destruction tests.
I had to find a way to destroy a ViewManagerConnection from the client lib.
In the new world, the ViewManagerSynchronizer is the InterfaceImpl, and lifetime of all other client objects hangs off it.
ViewManagerSynchronizer is owned by the ServiceConnector which is owned by the Application object, so I had to find a way to delete that. In the test, I simply reset the ServiceLoader by URL. This tears down the synchronizer & corresponding ViewManagerConnection, triggering the correct notifications.

Dave, I had to add a NULL check in the service manager code because it hadn't considered there could be a NULL loader for a URL. LMK if you have a different idea.

R=davemoore@chromium.org, sky@chromium.org
http://crbug.com/365012

Review URL: https://codereview.chromium.org/315863002