Convert format of pixels read from GL to BGRA on CPU if necessary
ReadPixels with format GL_BGRA_EXT is not necessarily supported by the
GLES2-based GPU command buffer. The support for this format needs to be
explicitly queried, and if it is not supported, the pixels need to be
converted from RGBA to BGRA on the CPU if a BGRA result is desired.

BUG=380129
TEST=WebGL conformance tests

Review URL: https://codereview.chromium.org/316553002