Cleanup: Remove the remaining instances of ChromiumTestShell from Telemetry scripts.
Chrome shell is still being referred as ChromiumTestShell in
some places, especially in the telemetry scripts. This
cleanup has dependency on changes in the buildbot
scripts as well where we are invoking the telemetry scripts
as ChromiumTestShell. The buildbot changes are uploaded as
part of https://crrev.com/302743002 to correct invocation as
ChromeShell instead ChromiumTestShell.

BUG=355730

Review URL: https://codereview.chromium.org/302653004