device/bluetooth: Fix bug in the GattServiceChanged event is sent.
This CL fixes a bug in which a BluetoothRemoteGattServiceChromeOS sent a
BluetoothGattService::Observer::GattServiceChanged event even when the
underlying D-Bus object that triggered the property change signal did not
correspond to the instance in question. This is fixed by correctly checking
that the object paths match.

BUG=none
R=keybuk@chromium.org

Review URL: https://codereview.chromium.org/313013003