TestPasswordManagerClient should inherit from StubPasswordManagerClient
This is a simple refactoring to cut down code duplication.
The test client in question provided a number of stubs for the
pure virtual methods of PasswordManagerClient, which is what
the StubPasswordManagerClient is for.

See also http://crbug.com/352566#c2, second part, for a description of this refactoring.

This CL makes the test client derive from the stub client (instead
from the client interface itself) to fix that.

BUG=352566

Review URL: https://codereview.chromium.org/316893002