Add more test for chrome.filesystem Drive support.
This CL adds test cases for multiple-file selection in the open dialog and
the save dialog (onto a new or existing file.)

The test code JS code is 100% shared with tests for local files
for checking the consistency. The test setup part introduces
the stuff specific to Drive.

BUG=223464

Review URL: https://codereview.chromium.org/315903002