Replace MockPasswordManagerDriver with StubPasswordManagerDriver.
Currently, there is a global MockPasswordManagerDriver, which mocks all
virtual methods of the PasswordManagerDriver interface. There is also
a locally (in a unit test) defined TestPasswordManagerDriver, which
inherits from the pure interface and provides stubs, a number of them
trivial.

That leads to code duplication and clutter.

This CL establishes the following structure:
It introduces StubPasswordManagerDriver, which is available to all
tests, and stubs all pure virtual methods of the driver interface.

Within the tests, mock or test drivers can inherit from the stub,
and only mock/modify methods interesting in that test.

See also http://crbug.com/352566#c2 for a description of this refactoring.

BUG=352566

Review URL: https://codereview.chromium.org/313113003