Call RemoveRoute() in EmbeddedWorkerDevToolsAgentHost::OnClientDetached().
From crrev.com/273741, DecrementWorkerRefCount() is called in ServiceWorkerProcessManager::Shutdown() to avoid a DCHECK in ProfileDestroyer::DestroyProfileWhenAppropriate while shutting down the browser.
But if once the inspector of ServiceWorker is opened, EmbeddedWorkerDevToolsAgentHost is added to the RenderProcessHost's listeners_.
So the RenderProcessHost will not be deleted in ServiceWorkerProcessManager::Shutdown() and this DCHECK fails whils shutting down.

In current code RemoveRoute() is called when the EmbeddedWorkerInstance is deleted.
But it is deleted after ProfileDestroyer::DestroyProfileWhenAppropriate is called.

To avoid this DCHECK failure I changed EmbeddedWorkerDevToolsManager to call RemoveRoute() in EmbeddedWorkerDevToolsAgentHost::OnClientDetached().
OnClientDetached is called when the inspector of ServiceWorker is closed.

BUG=368570,358657

Review URL: https://codereview.chromium.org/313553003