Added the "unmodifiable" property to chrome.bookmarks.BookmarkTreeNode.
This optional property indicates the reason why this BookmarkTreeNode can't
be modified by the user nor the extension. For now the only value defined is
"managed", indicating that this bookmark was configured by the system
administrator.

If the property is not present then the bookmark is modifiable (default).

BUG=49598
R=kalman@chromium.org

Review URL: https://codereview.chromium.org/320473002